Convert to a proper Python module
