+++ scraper/__main__.py
@@ -1,0 +1,3 @@
+from .scraper import scrapper
+
+scrapper()
