--- src/credentials.yaml
@@ -1,2 +1,0 @@
-email: email@email.fr
-password: my_plain_password
