--- src/input.txt
@@ -1,2 +1,0 @@
-https://www.facebook.com/andrew.ng.96
-https://www.facebook.com/zuck
